backend/requirements.txt
django>=5.0
django-ninja>=1.3
django-cors-headers>=4.3
google-generativeai>=0.8
openai>=1.40
pandas>=2.2
numpy>=1.26
python-dotenv>=1.0
gunicorn>=21.0
whitenoise>=6.6
